* 'main' of https://github.com/guluming/Spring_txt:
  Feat: "추가 txt파일을 읽어서 DB에 저장하는 기능"
  Feat: "추가 txt파일을 읽어서 DB에 저장하는 기능"

+++ src/main/java/com/example/spring_txt/controller/UserController.java
@@ -19,12 +19,12 @@ public class UserController {
     }
 
     @GetMapping("/user/id")
-    public ResponseEntity<Object> getUserId(@RequestParam("id") int id) {
+    public ResponseEntity<Object> getUserId(@RequestParam("id") Long id) {
         return userService.getUserId(id);
     }
 
     @DeleteMapping("/user/id")
-    public ResponseEntity<Object> deleteUserId(@RequestParam("id") int id) {
+    public ResponseEntity<Object> deleteUserId(@RequestParam("id") Long id) {
         return userService.deleteUserId(id);
     }
